apps/ln-tlp-poc/alice/requirements.txt
lnd-grpc-client==0.5.3
protobuf
protobuf_to_dict
gmpy2
psutil